dunders tests on both forms.
